[feat] Finish unwind.

+++ tapm-core/src/main/cpp/crash/memory_maps.cpp
@@ -112,7 +112,7 @@ bool tryFindAbortMsg(pid_t pid, LinkedList *maps, char *output) {
     }
 }
 
-bool findMemoryMapByAddress(uintptr_t address, LinkedList *maps, MemoryMap **target, MemoryMap ** previous) {
+bool findMemoryMapByAddress(uint64_t address, LinkedList *maps, MemoryMap **target, MemoryMap ** previous) {
     Iterator iterator;
     maps->iterator(&iterator);
     MemoryMap * p = nullptr;
